+++ ErsatzTV.Application/Troubleshooting/Queries/GetTroubleshootingSubtitles.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.Troubleshooting.Queries;
+
+public record GetTroubleshootingSubtitles(int MediaItemId) : IRequest<List<SubtitleViewModel>>;