+++ ErsatzTV.Core/FFmpeg/FFmpegStreamSelector.cs
@@ -1,4 +1,5 @@
-﻿using System.Diagnostics;
+﻿using System.Collections.Immutable;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using ErsatzTV.Core.Domain;
@@ -123,7 +124,7 @@ public class FFmpegStreamSelector : IFFmpegStreamSelector
     }
 
     public async Task<Option<Subtitle>> SelectSubtitleStream(
-        List<Subtitle> subtitles,
+        ImmutableList<Subtitle> subtitles,
         Channel channel,
         string preferredSubtitleLanguage,
         ChannelSubtitleMode subtitleMode)
@@ -140,6 +141,8 @@ public class FFmpegStreamSelector : IFFmpegStreamSelector
             return None;
         }
 
+        var candidateSubtitles = subtitles.ToList();
+
         bool useEmbeddedSubtitles = await _configElementRepository
             .GetValue<bool>(ConfigElementKey.FFmpegUseEmbeddedSubtitles)
             .IfNoneAsync(true);
@@ -147,10 +150,10 @@ public class FFmpegStreamSelector : IFFmpegStreamSelector
         if (!useEmbeddedSubtitles)
         {
             _logger.LogDebug("Ignoring embedded subtitles for channel {Number}", channel.Number);
-            subtitles = subtitles.Filter(s => s.SubtitleKind is not SubtitleKind.Embedded).ToList();
+            candidateSubtitles = candidateSubtitles.Filter(s => s.SubtitleKind is not SubtitleKind.Embedded).ToList();
         }
 
-        foreach (Subtitle subtitle in subtitles.Filter(s => s.SubtitleKind is SubtitleKind.Embedded && !s.IsImage)
+        foreach (Subtitle subtitle in candidateSubtitles.Filter(s => s.SubtitleKind is SubtitleKind.Embedded && !s.IsImage)
                      .ToList())
         {
             if (subtitle.IsExtracted == false)
@@ -159,7 +162,7 @@ public class FFmpegStreamSelector : IFFmpegStreamSelector
                     "Ignoring embedded subtitle with index {Index} that has not been extracted",
                     subtitle.StreamIndex);
 
-                subtitles.Remove(subtitle);
+                candidateSubtitles.Remove(subtitle);
             }
             else if (string.IsNullOrWhiteSpace(subtitle.Path))
             {
@@ -167,7 +170,7 @@ public class FFmpegStreamSelector : IFFmpegStreamSelector
                     "BUG: ignoring embedded subtitle with index {Index} that is missing a path",
                     subtitle.StreamIndex);
 
-                subtitles.Remove(subtitle);
+                candidateSubtitles.Remove(subtitle);
             }
         }
 
@@ -187,26 +190,26 @@ public class FFmpegStreamSelector : IFFmpegStreamSelector
                 _logger.LogDebug("Preferred subtitle language has multiple codes {Codes}", allCodes);
             }
 
-            subtitles = subtitles
+            candidateSubtitles = candidateSubtitles
                 .Filter(s => allCodes.Any(c => string.Equals(s.Language, c, StringComparison.OrdinalIgnoreCase)))
                 .ToList();
         }
 
-        if (subtitles.Count > 0)
+        if (candidateSubtitles.Count > 0)
         {
             Option<Subtitle> maybeSelectedSubtitle = subtitleMode switch
             {
-                ChannelSubtitleMode.Forced => subtitles
+                ChannelSubtitleMode.Forced => candidateSubtitles
                     .OrderBy(s => s.StreamIndex)
                     .Find(s => s.Forced)
                     .HeadOrNone(),
 
-                ChannelSubtitleMode.Default => subtitles
+                ChannelSubtitleMode.Default => candidateSubtitles
                     .OrderBy(s => s.Default ? 0 : 1)
                     .ThenBy(s => s.StreamIndex)
                     .HeadOrNone(),
 
-                ChannelSubtitleMode.Any => subtitles
+                ChannelSubtitleMode.Any => candidateSubtitles
                     .OrderBy(s => s.StreamIndex)
                     .HeadOrNone(),