+++ ErsatzTV.Core/Interfaces/FFmpeg/IFFmpegStreamSelector.cs
@@ -1,4 +1,5 @@
-﻿using ErsatzTV.Core.Domain;
+﻿using System.Collections.Immutable;
+using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.FFmpeg;
 
 namespace ErsatzTV.Core.Interfaces.FFmpeg;
@@ -15,7 +16,7 @@ public interface IFFmpegStreamSelector
         string preferredAudioTitle);
 
     Task<Option<Subtitle>> SelectSubtitleStream(
-        List<Subtitle> subtitles,
+        ImmutableList<Subtitle> subtitles,
         Channel channel,
         string preferredSubtitleLanguage,
         ChannelSubtitleMode subtitleMode);